+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Handle dates before 1900 during indexing [#2256](https://github.com/opendatateam/udata/pull/2256)
 - `spatial load` command is more resilient: make use of a temporary collection when `--drop` option is provided (avoid downtime during the load), in case of exception or keybord interrupt, temporary files and collections are cleaned up [#2261](https://github.com/opendatateam/udata/pull/2261)
 - Configurable Elasticsearch timeouts. Introduce `ELASTICSEARCH_TIMEOUT` as default/read timeout and `ELASTICSEARCH_INDEX_TIMEOUT` as indexing/write timeout [#2265](https://github.com/opendatateam/udata/pull/2265)
+- OEmbed support for organizations [#2273](https://github.com/opendatateam/udata/pull/2273)
 
 ## 1.6.13 (2019-07-11)